Fix for default (all) make target.
The 'Tests' target includes TestProxyCs so this doesn't need to be listed.
TestDeviceCs was a typo, of TestDvDeviceCs which doesn't build yet for Linux so this is removed too.  When TestDvDeviceCs does build for Linux it'll be added to the Tests target
Please enter the c mmit message for your changes. Lines starting