Add verbose logging option to network code.
Added Socket::LogVerbose() which logs all data sent or received from a socket.
Data can be logged as plain (ascii) text or hex.
This was easier to implement if we changed the static bsd-style functions to protected Socket functions.
Adding this removed the need for the #if 0'd code in the two Os.c files.  More importantly, it makes logs easier to interpret as we can now be selective about which sockets to log rather than always needing to wade through msearches and the like.