hanges to WebSocket server to support UPnP subscriptions.
The server now checks for sub-protocol upnp:event during the handshake.
Read() assumes it'll receive xml messages and checks for Subscribe/Renew/Unsubscribe methods.
Subscribe and Renew messages receive a SubscriptionTimeout response.
See the Zapp wiki for detalis of the format of the xml messages.

Refactored PropertyWriter (formerly in DviServerUpnp) to allow property writing code to be shared between the original s
erver and the WebSocket one.

Updated TestDvLights to set a mime type for html files as Chrome seems to require this.