Fix for warnings on Linux build.
The light dimming service was explicitly setting a 0..MAX_UINT range for a ui4 variable, leading to warnings that C90 and C99 would treat the type of 0xffffffff differently.  Since this only happens for a single variable in a single service, hide the problem by removing the allowedValues block from the service xml.  If we start to see more instances of the warning, we could consider changing the control point c++ templates to output something like (TUint)[num]ul.

While we're down to no warnings, change builds to use -Werror for consistency with Windows.