New approach to running the Js tests.
Added TestDeviceFinder which writes the presentation url for devices with a given service to stdout.
Js tests should first run this then run Chrome passing the presentation url.
There is now no need to specify a port to run the device stack server on so remove this from TesstDvTestBasic and the core stack.
Also added a -l switch for TestDvTestBasic allowing it to run on loopback only