Fix issue that was causing invocation tests to show 0 successful invocations for most DSes.
ReaderHttpChunked had an undocumented assumption that its underlying reader would have a buffer at least as big as any chunk of data.  This was clearly a bit mad.  Fixed by changing it to read in 4k blocks and documenting the assumption that clients constructed the reader with a buffer capable of reading 4k then updating InvocationUpnp to supply such a 4k buffer.