Fix for deadlock between CpiSubscription and CpProxy.
It was possible for one thread to hold the proxy property update locks and another to hold the subscription lock (which was taken before all EventUpdate functions).  Fixed this by holding a single subscription lock around all 3 EventUpdate calls.

Also replaced many guaranteed short uses of CpiSubscription::iLock with Stack.Mutex().  This seems like the right thing and was necessary to get the number of deadlocking threads above down to a comprehensible number!