Speculative fix for recent regular failures of tests on linux desktop.
I'd guess that the only way to fail the test with a NetworkError exception and not log the file/line that generated it is for the main thread to have thrown.  Code inspection suggests that Socket::Interrupt() is the most likely culprit for this.  It isn't helpful to allow this to fail (many destructors use it and there is never any remedial action to it failing) so log and error it encounters but don't propogate the error.