Change the WebSocket server to run on an OS-assigned port by default.
Preserve compatibility with current clients by adding APIs to hard-code a port and updated js test to set this to the old value.  Preparing the way for this to change in future, added an OhNet service which will always be published by the device stack and can be used to query the ws port.