Fix device list tests to wait the appropriate time for a refresh.
The tests were previously waiting for the mx time specified on the command line.  This defaulted to 0 (meaning use the stack's default) which was causing us to exit the tests too quickly.
Also fixed the fact that tests waited 2x mx for a refresh.  So the tests would actually take longer if your specified a low mx value on the command line!

See #38.