Provide implementation of niceness setting, when ATTEMPT_THREAD_NICENESS is defined.
Off by default, obviously.

According to Posix, niceness is a per-process property, not per-thread. However,
pthreads in Linux can, and does, apply nicess as a per-thread property.

We try and exploit that here by mapping our priority range onto a range of
nice-levels that do not require privaledges. In practise this means that we can
only really demote threads, not promote them.