Fix inline assembly for MLA (mult-accumulate) macro under FPM_PPC / OPT_ACCURACY
Input operands were marked as write-only ('=' modifier). This resulted in register
scheduling that accidentally hoses the hi-word of the accumulator:

  mulhw  r30,r27,r28
  mullw  r31,r27,r28
  addc   r25,r23,r31  -> shouldn't write to r25...
  adde   r24,r25,r30  -> because it's used for something else here!!!

Marking the output fields as 'read/write' ('+' modifier) seems to resolve this.

There might be an argument for adding 'earlyclobber' ('&' modifier) to the
output fields. More thought required, but TestCodec passes now, so we can take
it from there. :)

TC