Add IExitHandler, and global pointer gExitHandler.
All static error handling methods now notify the global exit handler,
gExitHandler, should it exist. libplatform can then be notified of
an exit condition, and have full access to appropriate data for the
CrashStore.

There's one hole in this implementation: TestFramework wants to install
an assert handler that throws AssertionFailed, and all unit tests rely
on this. The solution may be to kill SetAssertHandler() and replace it
with something more explicit, like SetAssertThrows(TBool)