Remove ability to change AssertHandler.
AFAICT, the only client of this was TestFramework. In order to support
libplatform's requirement to hook into ohNet error handling, it made sense
to narrow the ability of client code to alter assertion-handling
behaviour.

The previous API allowed client code to completely override the assert
handler, which conflicts directly with the IExitHandler strategy.