Changes to stream reader interface.
Read() now returns up to the number of bytes requested (was previously
exactly the number or an exception).  This allows Snaffle() and
ReadRemaining() functions to be removed.

ReadUntil() has been removed from IReader.  This is now available from a
new class ReaderUntil.

The few clients who wanted the old behaviour of Read() can call
ReadReplace() on ReaderUntil or ReaderProtocol::Read().