WebAppFramework: Don't instantiate iServer in ctor only to replace it.
This was creating some trouble with libosa-0.0.212 (CORE_LOCKING).

TestWebAppFramework was experiencing issues with TestHelperHttpReader::Connect()
returning code=0 after timing out.

Reinstantiating a SocketTcpServer on the same port/interface should work, but might
not be attempted in many places, so there may still be an issue in libosa. Let's
mask it for now and get on with things.