Apply HLS fix from Volkano 1: HttpReader: Set receive timeout in case server fails to close connection.
We had trouble with HLS radio streams: The supply thread would hang
indefinitely,
and it appeared to be stuck in HttpReader::Read().

Although HttpReader has a default connection timeout, it had no default
receive
timeout, which means that it can infact block indefinitely if the server
decides
to stop sending data.

We add a receive timeout of 10s. See
http://eng.linn.co.uk/trac/changeset/33929