[iOS/POSIX] Prevent accept() constantly looping and eating CPU time
Previously we would be calling accept() on a non-blocking socket. This
would always return EWOULDBLOCK / EAGAIN until there was a pending
connection.

This resulted in us always calling accept() select() throw in a tight
loop and ate a lot of CPU time. iOS has a sensitive runtime and if an
app uses too much CPU time a certain periods, like going in/out of
sleep, then it'll be killed off. The looping around here was causing
this to happen as of 1.24 or later.

Fix is prevent the socket being set as non-blocking, call select() first
to check for pending work before finally calling accept()