Build: Temp revert of minphoneos=12.0 back to minphoneos5.0
The change to target minimum of iOS 12.0 is incompatible with our toolchain when
linked with XCode 14.0+. Suspect this is to do with the older version of XCode
running on the Jenkins machine.

This allows us to build and deploy a new verison of ohNet compatible with CP apps
while we upgrade build systems.