Xml: Catch & rethrow XmlErrors to prevent iOS optimisation crashes
When an app is built with Xamarin iOS & XCode14, there appears to be a buggy
optmisation pass which prevents handling of XmlErrors when passing XML.

It was assumed that when parsing XML these exceptions would filter back up to
the caller. However, the newer toolchains appear to prevent this. Wrapping these
calls to rethrow the same exception appears to solve the problem while waiting
for Xamarin to solve the issue.

See GH-xamarinmacios-#16546