Exception: Fix buffer overflow getting thread names in exceptions
Thread names are limited to kMaxNameBytes. However, we then convert them
to a C-style string, which requires an additional byte for the
NULL-termination at the end. The longer thread names (mainly in Exakt)
could crash but then crash again when trying to get the log message as
we overflowed the name buffer.

This patch aims to address that by always including enough space for the
thread name AND the \0 at the end.