XmlFetch: Prevent use-after-free on repeated start/stop
Discovered during nightly tests that repeatedly starting and stopping
ohNet in short succession can cause various problems.

Firstly, it was discovered that XmlFetchManager and XmlFetcher threads
can overlap during shutdown whereby the termination of XmlFetchManager
attempts to interrupt the active XmlFetch in each fetcher. This XmlFetch
is not protected and so if we interrupt and then set to NULL before the
XmlFetcher thread is finished with the instance, we'll get a use-after-free
bug and crash.

Solution here is to put mutex protection around the XmlFetch instance to
prevent us from accessing the XmlFetch multiple times AND after it has
been freed.