CpiDevice: Attempt to prevent race condition on shutdown
Repeatedly starting and stopping ohNet can cause us to crash. This appears
to be the timer callback in CpiDeviceUpnp attempting to access it's member
reference to the parent device list AFTER that device list instance has been
deleted.

Therefore, the proposed solution is for CpiDevice to broadcast we're about
to be destoryed, giving you a chance to cancel timers / clear pending work.

See #8798